First real commit!
This is a small guide I'm puting together on how to build a PC.
This still in development. So when I reach pre-alpha value, I will notify.
This is PORTUGUESE ONLY at the time. There are currently many good
guides in english for building a PC.
Nonethless, I do plan on translating this guide to english.